+++ Assets/Blanku/Scripts/IllusionTeleport.cs
@@ -31,7 +31,7 @@ public class IllusionTeleport : MonoBehaviour {
 
 	}
 
-	public void CheckIllusion(int lookAngle){
+	public void CheckIllusion(float lookAngle){
 		Debug.Log (lookAngle);
 		portalOn = false;
