+++ Assets/Blanku/Scripts/LookAngleChangeEvent.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class LookAngleChangeEvent : UnityEvent<int> {
+public class LookAngleChangeEvent : UnityEvent<float> {
 	
 }
